Started working on the real allocator, still buggy
As the title says. It now instantiates objects on my own heap, but it
still crashes, currently working on that.

Fixed the myPayoad-naming bug in Object.hpp/cpp --> myPayload now :)

+++ Allocators/Allocator.cpp
@@ -40,7 +40,7 @@ void Allocator::swapHeaps() {
 void Allocator::freeAllSectors() {
 }
 
-int Allocator::gcAllocate(int size) {
+size_t Allocator::gcAllocate(int size) {
 	return -1;
 }
